Cleaning up grid view
