+++ src/components/fileBrowser/Grid/GridItemEmpty.vue
@@ -1,6 +1,6 @@
 <template>
   <p
-      class="                       px-6
+      class="px-6
                                     py-4
                                     whitespace-nowrap
                                     col-span-12
